refactor: move entities to related files
