+++ src/core/features/notes/utils.ts
@@ -1,0 +1,4 @@
+import { INoteContent } from '@core/features/notes/index';
+
+export const getNoteTitle = (note: INoteContent) =>
+	(note.title || note.text).slice(0, 25) || 'Empty note';
