+++ src/features/Workspace/index.tsx
@@ -1,5 +1,4 @@
 import React, { FC, PropsWithChildren, useEffect, useState } from 'react';
-import { ProfileContainer } from '@features/App';
 import {
 	attachmentsControllerContext,
 	filesRegistryContext,
@@ -8,6 +7,7 @@ import {
 } from '@features/Providers';
 import { SplashScreen } from '@features/SplashScreen';
 import { createNotesApi } from '@state/notes';
+import { ProfileContainer } from '@state/profiles';
 import { createTagsApi } from '@state/tags';
 import { createWorkspaceApi, useWorkspace } from '@state/workspace';
